changes the default meta data index lookup to use the UUID instead of the index name
this will make renaming of an index possible, otherwise the index will be deleted after the rename happens and the renamed index will have no shards at all.
by changing the lookup to use the UUID, the renamed index will not appear as deleted and so nothing will be deleted on the filesystem.