ec2-discovery: set current region if no endpoint available
The AwsEc2ServiceImpl instantiates a client and sets the endpoint which
can be specified explicitly by the ``ENDPOINT_SETTING`` or the
``REGION_SETTING``. If none of these settings are provided the client
would fall back to the endpoint of the ``us-east-1`` region.
This change however sets the client's region based on the region of the
running EC2 instance which can be obtained by an API call to the
instance metadata.

This behaviour was already patched in version 2.4.x of the Crate fork of
Elasticsearch and was accidentially removed with the ES5 upgrade.

(cherry picked from commit 9836e56fed31b65883ea1184b29526434309e567)