Patch IDEA support
This patch is necessary to avoid the ES workaround to be able to launch
tests from IntelliJ. The workaround forces all developers to disable the
IDEA launcher which has the additional drawback of limiting debugging
functionality like viewing stack dumps. The underlying issue is that
IntelliJ adds the same Jar twice to the classpath and the JarHell test
fails.

See https://github.com/elastic/elasticsearch/pull/24523/files